Different types of gaming mice: Which one suits you?
There are a lot of people that are looking forward to buy a good gaming mouse. A number of different models are available to you in the market. Many people think about the best gaming mouse considering the model but it is the type that plays a critical role. Here, we will consider different types of best gaming mice and take a detailed look at each of them.

All in one
The all purpose or all in one mouse is one of the most popular type of gaming mice available in the market. It comes in almost all mouse shapes and sizes. They are designed keeping in view all the purposes, from FPS to MMO to all other elements that play a role in gaming. If you are among individuals that play various games and need a mouse that can tackle with all of them then all purpose mouse should be your pick.

FPS
There are people that are into action games and love those precise shooting adventure ones like counter strike, call of duty, battlefield, etc. For such individuals, the FPS mice works the best. If you manage to optimize your game appropriately for the FPS play, you will be able to adjust your mouse in such a way that there will be a dedicated sniper button under your thumb. It will slow down DPI and assure that you can take difficult shots with ease. 

MMO
If you are into the multiplayer games like world of Warcraft, star wars, and other games from similar genre, the MMO mice can be of great help. These are usually bulky models that come with a list of buttons alongside. These will let you assign each button to a particular task and you can customize your game accordingly for perfect skills. For those of you that play games which require multiple tasks to be done one after the other in quick succession, assigning them to these buttons will be of great help.

RTS
Next up we have the RTS buttons and these are said to be the ones that are the pick of individuals who play real time strategy games. For individuals looking forward to a mouse that can back up their Star Craft II gameplay, the best pick can be an RTS button which is much like an all purpose mouse but with a little extra buttons to aid your gameplay.

Customizable
The customizable mice is costly and they totally worth the amount that you are going to spend (depending on your needs). A little hard to manage, they are highly customizable and you can adjust them according to your gameplay to assure that you use your mouse as a weapon at its best.